+++ cmd/clone.go
@@ -76,8 +76,10 @@ var cloneCmd = &cobra.Command{
 		format, _ := cmd.Flags().GetString("format")
 		noClosing, _ := cmd.Flags().GetBool("no-closing")
 		asJSON, _ := cmd.Flags().GetBool("json")
-		return newEntry(c,
+		return manageEntry(
+			c,
 			tec,
+			createTimeEntry(c),
 			viper.GetBool(INTERACTIVE),
 			viper.GetBool(ALLOW_PROJECT_NAME),
 			!noClosing,