+++ cmd/in.go
@@ -78,9 +78,10 @@ var inCmd = &cobra.Command{
 
 		format, _ := cmd.Flags().GetString("format")
 		asJSON, _ := cmd.Flags().GetBool("json")
-		return newEntry(
+		return manageEntry(
 			c,
 			tei,
+			createTimeEntry(c),
 			viper.GetBool(INTERACTIVE),
 			viper.GetBool(ALLOW_PROJECT_NAME),
 			true,