+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.18.0] - 2021-07-08
+
 ### Added
 
 - commands `in`, `clone` and `manual` will show a new "None" option on the projects list on the